refactor: rename services to utils, extract load_llm with Groq/Gemini fallback (#91)
